Modernize the plugin
Incomplete set of changes that may help a future maintainer of this plugin
as they modernize it and test it interactively for eventual release.

Require Jenkins 2.414.3 or newer.

Use most recent parent pom.

Compiles and passes tests with Java 21 on Linux.  Expected to compile
and pass tests with Java 17 on Windows.  Expected to compile and pass
tests with Java 11.

Powermock tests have been removed.  Powermock is dead and would need to
be replaced with Mockito.

Mockito tests have been removed.  Mockito references were outdated.
I'm not passionate enough about Mockito based tests to adapt them.
If a new maintainer wants to bring back those tests, they are in the
source history.

Selenium tests have been removed.  Jenkins user interface has changed
significantly enough since this plugin was last maintained that I expect
all the Selenium tests will fail for reasons that are unrelated to
plugin functionality.  I'm not passionate enough for Selenium testing to
do the necessary work to adapt the tests to the current Jenkins behavior.
If a new maintainer wants to bring back those tests, they are in the
source history.